+++ siibra/features/tabular/__init__.py
@@ -22,3 +22,4 @@ from .layerwise_cell_density import LayerwiseCellDensity
 from .receptor_density_fingerprint import ReceptorDensityFingerprint
 from .receptor_density_profile import ReceptorDensityProfile
 from .regional_timeseries_activity import RegionalBOLD
+from .cortical_profile import CorticalProfileSample
